ingestion/requirements.txt
scrapy==2.11.1
trafilatura==1.9.0
psycopg2-binary==2.9.9
w3lib==2.1.2
python-dateutil==2.9.0.post0
